requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==23.1.1
Automat==22.10.0
certifi==2022.9.24
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==39.0.0
daphne==3.0.2
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
hiredis==2.1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isodate==0.6.1
lxml==4.9.1
Markdown==3.4.1
msgpack==1.0.4
Pillow==9.3.0
platformdirs==2.5.4
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-decouple==3.6
pytz==2022.6
redis==4.4.2
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.10.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
urllib3==1.26.13
whitenoise==6.2.0
zeep==4.2.1
zope.interface==5.5.2
